working report + user system

+++ src/main/java/com/ntu/shvydkov/emerancy_bot/conditions/BotConditionObserved.java
@@ -2,8 +2,8 @@ package com.ntu.shvydkov.emerancy_bot.conditions;
 
 public interface BotConditionObserved {
 
-    BotCondition getCurrentBotConditionForUserById(Integer userId);
+    BotCondition getCurrentBotConditionForUserById(Long userId);
 
-    void setCurrentBotConditionForUserWithId(Integer userId, BotCondition botCondition);
+    void setCurrentBotConditionForUserWithId(Long userId, BotCondition botCondition);
 
 }
